requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
datasets==4.4.1
diffusers @ git+https://github.com/huggingface/diffusers@e6d46123091afd58281dc7487c0f6b67055683b9
dill==0.4.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@1ae4d917ed3badbdb1ffc167e0529f5a6d3c080d
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
